+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ Release Notes
 **New**
 
 - A... public constructor for Blob. I know, I'm sorry.
+- RowModel.isDirty. Updating non dirty models does not touch the database.
 
 
 ## v0.2.0